algutils/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "algutils"
version = "0.1.0"
description = "Algorithmic utilities: data tools, I/O, parameters, paths, caching, and more"
readme = "README.md"
requires-python = ">=3.9,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=1.10,<2.0",  # fixed_pydantic_yaml is written for pydantic v1
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "pyyaml>=5.4.0",
    "ruamel.yaml>=0.15.0",
    "regex>=2021.0.0",
    "tqdm>=4.60.0",
    "frozendict>=2.0.0",
    "python-box>=6.0.0",
    "joblib>=1.0.0",
    "configargparse>=1.0.0",
    "deprecated>=1.2.0",
    "semver>=2.13.0",
    "uncertainties>=3.1.0",
    "pint>=0.20.0",
    "zfpy",
    "xxhash>=3.0.0",
    "toolz>=0.11.0",
    "numba>=0.55.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
# Optional runtime extras
io = [
    "pillow>=8.0.0",
    "scikit-image>=0.19.0",
]
image = [
    "xxhash>=3.0.0",
    "toolz>=0.11.0",
]
math = [
    "numba>=0.55.0",
    "scikit-learn>=1.0.0",
]
units = [
    "pint>=0.20.0",
]
all = [
    "algutils[dev,io,image,math,units]",
]

[project.urls]
Homepage = "https://github.com/yourusername/algutils"
Repository = "https://github.com/yourusername/algutils"
Issues = "https://github.com/yourusername/algutils/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
algutils = [
    "**/*.yml",
    "**/*.yaml",
    "**/*.json",
    "**/*.md",
    "**/*.ipynb",
    "**/*.puml",
    "**/*.jscm",
    "**/*.org",
    "py.typed",
]

[tool.pytest.ini_options]
testpaths = ["src/algutils"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
